+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/AdminOAuthProvider.kt
@@ -94,7 +94,7 @@ data class AdminOAuthProvider (
 
     /* The oauth provider id. */
     @SerializedName("id")
-    val id: kotlin.Double,
+    val id: kotlin.Int,
 
     @SerializedName("updated_at")
     val updatedAt: kotlin.String? = null
