+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/CommunityIdQueryI.kt
@@ -29,7 +29,7 @@ data class CommunityIdQueryI (
 
     /* The community id. */
     @SerializedName("id")
-    val id: kotlin.Double
+    val id: kotlin.Int
 
 ) {
