+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/CreateCommentLike.kt
@@ -30,7 +30,7 @@ data class CreateCommentLike (
 
     /* The comment id. */
     @SerializedName("comment_id")
-    val commentId: kotlin.Double,
+    val commentId: kotlin.Int,
 
     /* True means Upvote, False means Downvote, and None means remove vote. */
     @SerializedName("is_upvote")
