+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/CreateCommunityReport.kt
@@ -33,7 +33,7 @@ data class CreateCommunityReport (
 
     /* The community id. */
     @SerializedName("community_id")
-    val communityId: kotlin.Double
+    val communityId: kotlin.Int
 
 ) {
