+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/CreateCommunityTag.kt
@@ -37,7 +37,7 @@ data class CreateCommunityTag (
 
     /* The community id. */
     @SerializedName("community_id")
-    val communityId: kotlin.Double,
+    val communityId: kotlin.Int,
 
     @SerializedName("color")
     val color: TagColor? = null,
