+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/DeleteOAuthProvider.kt
@@ -29,7 +29,7 @@ data class DeleteOAuthProvider (
 
     /* The oauth provider id. */
     @SerializedName("id")
-    val id: kotlin.Double
+    val id: kotlin.Int
 
 ) {
