+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/DeleteTagline.kt
@@ -29,7 +29,7 @@ data class DeleteTagline (
 
     /* The tagline id. */
     @SerializedName("id")
-    val id: kotlin.Double
+    val id: kotlin.Int
 
 ) {
