+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/DistinguishComment.kt
@@ -33,7 +33,7 @@ data class DistinguishComment (
 
     /* The comment id. */
     @SerializedName("comment_id")
-    val commentId: kotlin.Double
+    val commentId: kotlin.Int
 
 ) {
