+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/EditCommunity.kt
@@ -37,7 +37,7 @@ data class EditCommunity (
 
     /* The community id. */
     @SerializedName("community_id")
-    val communityId: kotlin.Double,
+    val communityId: kotlin.Int,
 
     @SerializedName("visibility")
     val visibility: CommunityVisibility? = null,
