+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/EditCommunityNotifications.kt
@@ -34,7 +34,7 @@ data class EditCommunityNotifications (
 
     /* The community id. */
     @SerializedName("community_id")
-    val communityId: kotlin.Double
+    val communityId: kotlin.Int
 
 ) {
