+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/EditOAuthProvider.kt
@@ -40,7 +40,7 @@ data class EditOAuthProvider (
 
     /* The oauth provider id. */
     @SerializedName("id")
-    val id: kotlin.Double,
+    val id: kotlin.Int,
 
     @SerializedName("enabled")
     val enabled: kotlin.Boolean? = null,
