+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/EditPostNotifications.kt
@@ -34,7 +34,7 @@ data class EditPostNotifications (
 
     /* The post id. */
     @SerializedName("post_id")
-    val postId: kotlin.Double
+    val postId: kotlin.Int
 
 ) {
