+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/EditPrivateMessage.kt
@@ -33,7 +33,7 @@ data class EditPrivateMessage (
 
     /* The private message id. */
     @SerializedName("private_message_id")
-    val privateMessageId: kotlin.Double
+    val privateMessageId: kotlin.Int
 
 ) {
