+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/EditTagline.kt
@@ -33,7 +33,7 @@ data class EditTagline (
 
     /* The tagline id. */
     @SerializedName("id")
-    val id: kotlin.Double
+    val id: kotlin.Int
 
 ) {
