+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/FollowMultiCommunity.kt
@@ -32,7 +32,7 @@ data class FollowMultiCommunity (
     val follow: kotlin.Boolean,
 
     @SerializedName("multi_community_id")
-    val multiCommunityId: kotlin.Double
+    val multiCommunityId: kotlin.Int
 
 ) {
