+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/GetCommentI.kt
@@ -29,7 +29,7 @@ data class GetCommentI (
 
     /* The comment id. */
     @SerializedName("id")
-    val id: kotlin.Double
+    val id: kotlin.Int
 
 ) {
