+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/GetCommunityI.kt
@@ -34,7 +34,7 @@ data class GetCommunityI (
 
     /* The community id. */
     @SerializedName("id")
-    val id: kotlin.Double? = null
+    val id: kotlin.Int? = null
 
 ) {
