+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/GetPersonDetailsI.kt
@@ -34,7 +34,7 @@ data class GetPersonDetailsI (
 
     /* The person id. */
     @SerializedName("person_id")
-    val personId: kotlin.Double? = null
+    val personId: kotlin.Long? = null
 
 ) {
