+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/GetRegistrationApplicationI.kt
@@ -29,7 +29,7 @@ data class GetRegistrationApplicationI (
 
     /* The person id. */
     @SerializedName("person_id")
-    val personId: kotlin.Double
+    val personId: kotlin.Long
 
 ) {
