+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/HidePost.kt
@@ -33,7 +33,7 @@ data class HidePost (
 
     /* The post id. */
     @SerializedName("post_id")
-    val postId: kotlin.Double
+    val postId: kotlin.Int
 
 ) {
