+++ api/src/main/kotlin/com/idunnololz/summit/api/dto/lemmy/v4/models/LockComment.kt
@@ -37,7 +37,7 @@ data class LockComment (
 
     /* The comment id. */
     @SerializedName("comment_id")
-    val commentId: kotlin.Double
+    val commentId: kotlin.Int
 
 ) {
